requirements.txt
flask==2.3.3
gunicorn==21.2.0
flask-cors==4.0.0
python-telegram-bot==20.7
gdown==4.7.1
requests==2.31.0